prevent crash in remove liquidity if wrong token in url
