+++ src/contexts/index.js
@@ -8,6 +8,7 @@ import { WalletContext } from './WalletContext';
 import { NotificationContext } from './NotificationContext';
 import { ApplicationContext } from './ApplicationContext';
 import { GameEditionContext } from './GameEditionContext';
+import { LiquidityContext } from './LiquidityContext';
 
 export function useKaddexWalletContext() {
   return useContext(KaddexWalletContext);
@@ -35,4 +36,7 @@ export function useApplicationContext() {
 }
 export function useGameEditionContext() {
   return useContext(GameEditionContext);
+}
+export function useLiquidityContext() {
+  return useContext(LiquidityContext);
 }
